+++ src/briefy/leica/models/dashboard/pm.py
@@ -35,7 +35,7 @@ delivered_orders_pm = select([
                 Order.state == 'delivered',
                 Order.accept_date.is_(None)
             ), 1)], else_=0)
-    ).label('newly-delivered'),
+    ).label('newly_delivered'),
     func.sum(
         case([(
             and_(
@@ -49,7 +49,7 @@ delivered_orders_pm = select([
                 Order.state == 'delivered',
                 Order.accept_date.isnot(None)
             ), 1)], else_=0)
-    ).label('re-delivered'),
+    ).label('re_delivered'),
     func.sum(
         case([(Order.state == 'accepted', 1)], else_=0)
     ).label('completed'),