+++ tests/views/test_leadorders_view.py
@@ -8,8 +8,8 @@ import pytest
 
 
 @pytest.mark.usefixtures('create_dependencies')
-class TestOrderView(BaseVersionedTestView):
-    """Test OrderService view."""
+class TestLeadOrderView(BaseVersionedTestView):
+    """Test LeadOrderService view."""
 
     base_path = '/orders'
     dependencies = [
@@ -22,13 +22,13 @@ class TestOrderView(BaseVersionedTestView):
         'customer_user', 'project_manager', 'scout_manager', 'location',
         'external_id', 'assignment', 'assignments', 'price', 'versions'
     ]
-    file_path = 'data/orders.json'
-    model = models.Order
-    initial_wf_state = 'received'
+    file_path = 'data/leadorders.json'
+    model = models.LeadOrder
+    initial_wf_state = 'new'
     UPDATE_SUCCESS_MESSAGE = ''
     NOT_FOUND_MESSAGE = ''
     update_map = {
-        'title': 'New Order Title changed!',
+        'title': 'New LeadOrder Title changed!',
         'price': 10000,
         'price_currency': 'USD',
         'availability': [
@@ -36,18 +36,3 @@ class TestOrderView(BaseVersionedTestView):
             (datetime_utcnow() + timedelta(days=21)).isoformat()
         ]
     }
-
-    def test_put_invalid_asset_tyoe(self, app, obj_payload):
-        """Asset type should match one of the possible values."""
-        payload = obj_payload
-        del(payload['availability'])
-        obj_id = payload['id']
-        payload['asset_types'] = ['Foobar']
-        request = app.put_json('{base}/{id}'.format(base=self.base_path, id=obj_id),
-                               payload, headers=self.headers, status=400)
-        result = request.json
-        error = result['errors'][0]
-        assert result['status'] == 'error'
-        assert error['name'] == 'asset_types'
-        assert error['location'] == 'body'
-        assert 'Invalid type of asset' in error['description']